Update to 5.0-8987 for the September and October 2018 progress report
I have confirmed through personal testing that the bluetooth passthrough feature does work with this release. However, it requires the user to manually create a udev rule in the host machine as detailed in the dolphin documentation: https://wiki.dolphin-emu.org/index.php?title=USB_Passthrough#Linux

This is hardly ideal, but it's the way upstream recommends, and flatpak doesn't currently offer a means to improve it.